Preparation for pattern-meta training.

+++ Matcher.h
@@ -6,27 +6,23 @@
 #include <bitset>
 
 class Matcher {
-public:    
-    Matcher();    
-    
-    //static const int UNITY = 2;
-    //static const int THRESHOLD = 2;
-    //static const int PROXFACTOR = 40;
+public:
+    Matcher();
 
-    uint16 matches(int color, int pattern);
+    float matches(int color, int pattern);
 
     /*
         return the "global" matcher
-    */            
+    */
     static Matcher* get_Matcher(void);
     static void set_Matcher(Matcher * m);
 
-private:            
+private:
     static unsigned short clip(double val);
 
-    static Matcher* s_matcher;   
+    static Matcher* s_matcher;
 
-    std::tr1::array<std::vector<uint16>, 2> m_patterns;
+    std::tr1::array<std::vector<float>, 2> m_patterns;
 };
 
 #endif
